core-service/pyproject.toml
[project]
name = "core-service"
version = "0.1.0"
description = "Core microservice"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
]
